requirements.txt
# Core dependencies for Nikola Tesla Chat
python-dotenv==1.2.1
langchain-google-genai==3.1.0
gradio==6.0.0
